keep stuff updated

+++ commands/util/clean.js
@@ -1,3 +1,8 @@
+const {
+  Tools,
+} = require('../../tools');
+const tools = new Tools();
+
 module.exports = {
   name: 'clean',
   usage: '[user|all] [number]',
@@ -17,27 +22,27 @@ module.exports = {
     }
 
     const canDelOth = message.channel.permissionsFor(message.member).has("MANAGE_MESSAGES");
-    const hasAdmin = message.channel.permissionsFor(message.author).has("ADMINISTRATOR");
+    const hasAdmin = message.channel.permissionsFor(message.member).has("ADMINISTRATOR");
     if (!canDelOth) {
-      return message.channel.send(message.author + " You don't have permission to use this command! " + weirdChamp);
+      return message.channel.send(`${message.author} You don't have permission to use this command! ${weirdChamp}`);
     }
 
     let amnt;
     if (hasTwoArgs) {
       amnt = parseInt(args[1]);
       if (isNaN(amnt)) {
-        return message.channel.send(message.author + ' Please use numbers!');
+        return message.channel.send(`${message.author} Please use numbers!`);
       }
     } else {
       amnt = 100;
     }
     if (amnt > 100) {
-      return message.channel.send(message.author + ' Please use a number from 1-100!');
+      return message.channel.send(`${message.author} Please use a number from 1-100!`);
     }
     if (args.length === 0) {
       message.delete();
-      return message.channel.fetchMessages().then(messages => {
-        if (amnt > messages.array().size) return message.channel.send(message.author + ' Could not find that many messages!');
+      return message.channel.messages.fetch().then(messages => {
+        if (amnt > messages.array().size) return message.channel.send(`${message.author} Could not find that many messages!`);
         const authMessages = messages.filter(msg => {
           return msg.author.equals(message.author);
         });
@@ -50,13 +55,13 @@ module.exports = {
     }
 
     if (args[0] === "all") {
-      if (!hasAdmin) return message.channel.send(message.author + "You don't have perission to use this command! " + weirdChamp);
+      if (!hasAdmin) return message.channel.send(`${message.author} You don't have perission to use this command! ${weirdChamp}`);
       message.delete();
       try {
-        return message.channel.fetchMessages({
+        return message.channel.messages.fetch({
           limit: amnt,
         }).then(messages => {
-          if (amnt > messages.array().size) return message.channel.send(message.author + ' Could not find that many messages!');
+          if (amnt > messages.array().size) return message.channel.send(`${message.author}` + ' Could not find that many messages!');
           message.channel.bulkDelete(messages).then((deleted) => {
             message.channel.send('Deleted ' + deleted.size + ' messages from this channel!').then(async sent => {
               awaitDelete(sent);
@@ -67,15 +72,15 @@ module.exports = {
         return message.channel.send('There were no messages found in this channel!');
       }
     }
-    const mention = message.mentions.users.first();
+    const mention = tools.parseMention(args[0], client);
     if (!mention) {
       return message.channel.send('That user does not exist!');
     } else {
       message.delete();
-      return message.channel.fetchMessages({
+      return message.channel.messages.fetch({
         limit: amnt,
       }).then(messages => {
-        if (amnt > messages.array().size) return message.channel.send(message.author + ' Could not find that many messages!');
+        if (amnt > messages.array().size) return message.channel.send(`${message.author} Could not find that many messages!`);
         const mentionMessages = messages.filter(msg => {
           return msg.author.equals(mention);
         });
@@ -92,15 +97,12 @@ module.exports = {
 //delete emote
 async function awaitDelete(sent) {
   sent.react("❎");
-  let emotes = 0;
-  sent.awaitReactions(reaction => {
+  sent.awaitReactions((reaction, user) => {
+    if (user.bot) return;
     if (reaction.emoji.name === "❎") {
-      emotes++;
+      sent.delete();
     }
-    switch (emotes) {
-      case 2:
-        sent.delete();
-        break;
-    }
+  }, {
+    time: 20000,
   });
 }
